requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
python-jose
pydantic
httpx
pytest
pytest-asyncio
pytest-cov
black
mypy
flake8
bandit
isort
ruff
